information collector/requirements.txt
streamlit==1.16.0
flask==2.2.5
pymongo==4.5.0
requests==2.31.0
python-dotenv==1.0.0
